fix: remove unnecessary DefaultEventLoopGroup tests and fix as reviewed
